controlli per autore del post

+++ app/Http/Controllers/Admin/PostController.php
@@ -18,8 +18,8 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts= Post::all();
         $user= Auth::user();
+        $posts= $user->posts;
         return view('admin.posts.index', compact('posts'));
     }
 
@@ -57,6 +57,7 @@ class PostController extends Controller
 
         $newPost= New Post();
         $newPost->fill($data);
+        $newPost->user_id=Auth::user();
         $newPost->save();
         
         if(isset($data['tags'])) {
@@ -73,6 +74,9 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post->user_id!==Auth::user()){
+            abort(403);
+        }
         return view('admin.posts.show', compact('post'));
     }
 
@@ -84,6 +88,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+
+        if($post->user_id!==Auth::user()){
+            abort(403);
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         $postTags = $post->tags->map(function ($item) {
@@ -102,6 +111,10 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id!==Auth::user()){
+            abort(403);
+        }
+
         $request->validate([
             'title' => 'required|max:255',
             'content' => 'required|max:65535',
@@ -130,6 +143,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id!==Auth::user()){
+            abort(403);
+        }
         $post->delete();
         return redirect()->route('admin.posts.index');
     }
